Do not specify patch version of Docker images
This makes `docker-compose pull` automatically update locally cached
images to the latest patch version with all security fixes making
keeping Docker images up-to-date technically simpler (this is related to
[Issue #124](https://github.com/ise621/metabase/issues/124)). However,
it makes the setup less reproducible. It may happen that a new patch
version breaks things though this should be unlikely.